requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.35.0
sqlalchemy==2.0.49
pymysql==1.1.2
python-dotenv==1.2.2
openai==1.83.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
langchain-community==0.4.1
langchain-text-splitters==1.1.1
langchain-huggingface==0.3.0
langchain-chroma==0.2.0
chromadb==1.5.5
sentence-transformers==5.3.0
dashscope==1.21.0